perf: rewrite lexer dispatch, skip trivia in cursor, shrink spans, exact-size AST lists (#27)
Lexing and parsing are 11-30% faster across all benchmarks and the parser
uses about half the arena memory, with byte-identical tokens and errors
verified against a golden dump of all 149 corpus fixtures.

Lexer:
- Dispatch each token from a 256-entry byte-class table instead of the
  per-byte state machine; scan strings, block strings and comments with
  memchr. Numbers keep the original state machine.
- Check the pending in-string error only on string paths (the only paths
  that can set it), with a debug_assert documenting the invariant.
- Fix a pre-existing panic: `.` followed by a multibyte character sliced
  the error data inside the character.

Parser:
- Fetch tokens via a crate-internal Lexer::next_significant() that skips
  whitespace/comment/comma trivia inside the cursor without materializing
  a Token/Result per trivia item. Token-limit counting is unchanged.
- Borrow string values from the source when they contain no escape
  sequences (or no `\r` for block strings) instead of allocating.
- Build AST lists in a single reusable LIFO scratch stack and copy them
  into the arena once at the exact final size: arena vecs cannot grow in
  place when element parsing allocates in between, so every doubling
  stranded a dead copy in the arena.

AST (breaking):
- Span offsets are now u32 (halves every span-carrying node, e.g.
  Selection 176 -> 144 bytes). Parser::new asserts input <= 4 GiB.

Benchmarks (criterion, vs saved baseline):
  query_lexer       495 ns  -> 347 ns   (-30%)
  supergraph_lexer  17.4 us -> 13.5 us  (-22%)
  supergraph_parser 39.9 us -> 29.7 us  (-26%)
  many_aliases      309 us  -> 274 us   (-11%)
  query_parser      975 ns  -> 920 ns   (-3..6%)

Memory (counting allocator + Allocator::used_bytes):
  supergraph arena  111 KB  -> 52 KB    (-53%), heap allocs 42 -> 4
  alias arena       1.17 MB -> 593 KB   (-49%), peak heap -60%

+++ crates/oxc_graphql_parser/src/ast.rs
@@ -5,14 +5,19 @@ use std::slice::Iter;
 
 pub use oxc_allocator::{Box as AstBox, Vec as AstVec};
 
+/// A half-open byte range into the source text.
+///
+/// Offsets are `u32`: source texts are limited to 4 GiB (asserted by
+/// [`crate::Parser::new`]), which halves the size of every AST node that
+/// carries a span.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Default, Hash)]
 pub struct Span {
-    pub start: usize,
-    pub end: usize,
+    pub start: u32,
+    pub end: u32,
 }
 
 impl Span {
-    pub fn new(start: usize, end: usize) -> Self {
+    pub fn new(start: u32, end: u32) -> Self {
         Self { start, end }
     }
 }
